restruct models (ProfileView)
