+++ MealBytes/View/Profile/ProfileView/Models/AlertContentProfile.swift
@@ -36,77 +36,3 @@ struct AlertContentProfile: Identifiable {
         self.isSuccess = isSuccess
     }
 }
-
-enum AlertTypeProfileView {
-    case signOut
-    case deleteAccount
-    case changePassword
-    
-    var title: String {
-        switch self {
-        case .signOut: return "Sign Out"
-        case .deleteAccount: return "Delete Account"
-        case .changePassword: return "Change Password"
-        }
-    }
-    
-    var destructiveTitle: String {
-        switch self {
-        case .signOut: return "Sign Out"
-        case .deleteAccount: return "Delete"
-        case .changePassword: return "Update Password"
-        }
-    }
-    
-    var defaultMessage: String {
-        switch self {
-        case .signOut:
-            return "Signing out will require signing in again to access the account."
-        case .deleteAccount:
-            return """
-            To delete the account, enter the password associated with it.
-            Data and account details will be permanently erased. This action cannot be undone.
-            """
-        case .changePassword:
-            return "Provide the current password and a new password to update account credentials."
-        }
-    }
-}
-
-enum ProfileMessage {
-    case passwordTooShort
-    case passwordMismatch
-    case incorrectPassword
-    case emailMissing
-    case passwordUpdateFailed
-    case passwordUpdateSuccess
-    
-    var text: String {
-        switch self {
-        case .passwordTooShort:
-            return """
-            Failed to update the password.
-            The password must be at least 6 characters long.
-            """
-        case .passwordMismatch:
-            return """
-            Failed to update the password.
-            New password and confirmation do not match.
-            """
-        case .incorrectPassword:
-            return """
-            The password entered is incorrect.
-            To delete the account, provide the correct password.
-            """
-        case .emailMissing:
-            return "Email is missing."
-        case .passwordUpdateFailed:
-            return """
-            Failed to update the password.
-            Check the current password and try again.
-            """
-        case .passwordUpdateSuccess:
-            return "Password has been successfully updated."
-        }
-    }
-}
